Decoupled crypto functions in main.py in its own file.
